+++ Gemfile
@@ -2,5 +2,5 @@ source "https://rubygems.org"
 
 git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
 
-# Specify your gem's dependencies in support_developer_github-base.gemspec
+# Specify your gem's dependencies in pghub-base.gemspec
 gemspec
