+++ lib/pghub/base/version.rb
@@ -1,4 +1,4 @@
-module SupportDeveloperGithub
+module PgHub
   module Base
     VERSION = "0.1.0"
   end
