+++ lib/pghub/github_api.rb
@@ -1,5 +1,5 @@
-require 'support_developer_github/github_api/issue'
-require 'support_developer_github/github_api/comment'
+require 'pghub/github_api/issue'
+require 'pghub/github_api/comment'
 
 class GithubAPI
   include Issue
