+++ lib/pghub/github_api/comment.rb
@@ -1,5 +1,5 @@
-require 'support_developer_github/config'
-require 'support_developer_github/github_api/connection'
+require 'pghub/config'
+require 'pghub/github_api/connection'
 
 class GithubAPI
   module Comment
@@ -7,7 +7,7 @@ class GithubAPI
 
     def post(content)
       connection.post do |req|
-        req.url "#{issue_path}/comments?access_token=#{SupportDeveloperGithub.config.github_access_token}"
+        req.url "#{issue_path}/comments?access_token=#{PgHub.config.github_access_token}"
         req.headers['Content-Type'] = 'application/json'
         req.body = { body: content }.to_json
       end
