--- lib/support_developer_github/engine.rb
@@ -1,11 +1,0 @@
-require 'support_developer_github/config'
-
-module SupportDeveloperGithub
-  def self.config
-    @config ||= SupportDeveloperGithub::Config.new
-  end
-
-  def self.configure(&block)
-    yield(config) if block_given?
-  end
-end
