+++ spec/pghub/base_spec.rb
@@ -1,8 +1,8 @@
 require "spec_helper"
 
-RSpec.describe SupportDeveloperGithub::Base do
+RSpec.describe PgHub::Base do
   it "has a version number" do
-    expect(SupportDeveloperGithub::Base::VERSION).not_to be nil
+    expect(PgHub::Base::VERSION).not_to be nil
   end
 
   it "does something useful" do
